// src/app/news/models/pagecontent.ts
export interface Pagecontent {
  title?: string;
  content: string;
  author?: string;
  date_published?: string;
  lead_image_url?: string;
  dek?: string;
  next_page_url?: string;
  domain?: string;
  excerpt: string;
  word_count: number;
  direction: string;
  total_pages: number;
  rendered_pages: number;
}